[ fixed #5065 ] by only reducing rhss with exact clauses
In #4560 we try to help the termination checker by getting rid of
recursive calls via reduction of the rhss with non-recursive clauses.
However, the implementation was naive in that it did not exclude
catch-all (= non-exact) clauses.

In this patch, we rectify this problem.  To this end, we need to
remember whether the coverage checker recognized a clause as exact.
Hence, we introduce a new field clauseExact to the internal clauses.
This is ordinarily initialized with Nothing until the coverage checker
computes its value.  However, in simple cases we can already set it
when the clause is created.  Examples are absurd clauses, which are
always non-exact, and "missing" and "cubical system" clauses which are
always exact (at least so I hope).

This patch also changes clauseCatchall for absurd clauses to True in
general.  Previously, only those created in Rules/Def.hs where True,
the other two (e.g. in Rules/Term.hs) where False.